techstore/pyproject.toml
[project]
name = "techstore"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "asgiref (==3.11.0)",
    "certifi (==2025.11.12)",
    "cffi (==2.0.0)",
    "charset-normalizer (==3.4.4)",
    "cryptography (==46.0.3)",
    "defusedxml (==0.7.1)",
    "django (==6.0)",
    "django-environ (==0.12.0)",
    "djangorestframework (==3.16.1)",
    "djangorestframework-simplejwt (==5.5.1)",
    "djoser (==2.3.3)",
    "drf-yasg (==1.21.11)",
    "idna (==3.11)",
    "inflection (==0.5.1)",
    "oauthlib (==3.3.1)",
    "packaging (==25.0)",
    "pillow (==12.0.0)",
    "psycopg2-binary (==2.9.11)",
    "pycparser (==2.23)",
    "pyjwt (==2.10.1)",
    "python3-openid (==3.2.0)",
    "pytz (==2025.2)",
    "pyyaml (==6.0.3)",
    "requests (==2.32.5)",
    "requests-oauthlib (==2.0.0)",
    "social-auth-app-django (==5.7.0)",
    "social-auth-core (==4.8.3)",
    "sqlparse (==0.5.4)",
    "uritemplate (==4.2.0)",
    "urllib3 (==2.6.2)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
